Factor network methods from APIResource to StripeResponseGetter
